IMPLEMENTATIONS I MADE ARE WORKING

+++ CKK.DB/Repository/ProductRepository.cs
@@ -32,31 +32,31 @@ namespace CKK.DB.Repository
 
 
 
-        public int Delete(int id)
-        {
-            var sql = "DELETE FROM Products WHERE Id = @Id";
-            using (var connection = _connectionFactory.GetConnection)
-            {
-                connection.Open();
-                return connection.Execute(sql, new { Id = id });
-            }
-        }
+        //public int Delete(int id)
+        //{
+        //    var sql = "DELETE FROM Products WHERE Id = @Id";
+        //    using (var connection = _connectionFactory.GetConnection)
+        //    {
+        //        connection.Open();
+        //        return connection.Execute(sql, new { Id = id });
+        //    }
+        //}
 
         //IMPLEMENTED NEW DELETE WORKS
-        //public int Delete(Product entity)
-        //{
+        public int Delete(Product entity)
+        {
 
-        //    if (entity != null)
-        //    {
-        //        var sql = "DELETE FROM Products WHERE Id = @Id";
-        //        using (var connection = _connectionFactory.GetConnection)
-        //        {
-        //            connection.Open();
-        //            return connection.Execute(sql, new { Id = entity.Id });
-        //        }
-        //    }
-        //    return 0;
-        //}
+            if (entity != null)
+            {
+                var sql = "DELETE FROM Products WHERE Id = @Id";
+                using (var connection = _connectionFactory.GetConnection)
+                {
+                    connection.Open();
+                    return connection.Execute(sql, new { Id = entity.Id });
+                }
+            }
+            return 0;
+        }
 
         public List<Product> GetAll()
         {
